Moved GetVersion function to util
- modified tests (manually verfied version in built executable)
- remove version flags in release script
