+++ release
@@ -26,7 +26,7 @@ do
   fi
   path="$name"
   log "building $path..."
-  CGO_ENABLED=0 GOOS=$os GOARCH=$arch go build -C .. -ldflags "-w -X github.com/NicksPatties/sweet/cmd/about.version=v0.1.0 -X github.com/NicksPatties/sweet/cmd/version.version=$version -X github.com/NicksPatties/sweet/cmd/about.version=$version" -o "$builds/$path"
+  CGO_ENABLED=0 GOOS=$os GOARCH=$arch go build -C .. -ldflags "-w -X github.com/NicksPatties/sweet/util.version=$version" -o "$builds/$path"
   log "$path built"
   assets+=("$path")
   log "generating checksum for $path..."
